Configuración de las rutas para proteger las que nos interesan
