+++ src/firebase/AuthContext.jsx
@@ -1,0 +1,24 @@
+import { Children, createContext, useContext, useEffect, useState } from "react";
+import { auth } from "./config";
+import { onAuthStateChanged } from "firebase/auth";
+
+const AuthContext = createContext();
+
+export const useAuth = () => useContext(AuthContext);
+
+export const AuthProvider = ({ children }) => {
+    const [currentUser, setCurrentUser] = useState(null);
+
+    useEffect(() => {
+        const unsubscribe = onAuthStateChanged(auth, (user) => {
+            setCurrentUser(user);
+        });
+        return () => unsubscribe();
+    }, []);
+
+    return (
+        <AuthContext.Provider value={{ currentUser }}>
+            {children}
+        </AuthContext.Provider>
+    )
+}
