+++ src/firebase/PrivateRoute.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+import { Navigate } from "react-router-dom";
+import { useAuth } from "./AuthContext";
+
+const PrivateRoute = ({ children }) => {
+    const { currentUser } = useAuth();
+
+    return currentUser ? children : <Navigate to="/login" />;
+};
+
+export default PrivateRoute;
